backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "traffic-viz-backend"
version = "0.0.1"
description = "FastAPI backend for traffic-viz"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.123.9",
    "uvicorn[standard]==0.38.0",
    "shapely==2.1.2",
    "pygeoops==0.6.0",
    "python-multipart==0.0.20",
    "pillow==11.3.0",
]

[project.optional-dependencies]
dev = [
    "fastapi[standard]==0.123.9",
    "jupyterlab>=4.0",
    "ipykernel>=6.0",
    "pyinstaller==6.18.0",
]
build = [
    "pyinstaller==6.18.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]